@@ -28,11 +28,11 @@
     "cleave.js": "0.7.15",
     "immutable": "3.8.1",
     "lodash": "4.17.4",
-    "react": "15.4.2",
+    "react": "15.5.0",
     "react-animations": "0.1.0",
     "react-dnd": "2.2.4",
     "react-dnd-html5-backend": "2.2.4",
-    "react-dom": "15.4.2",
+    "react-dom": "15.5.0",
     "react-fine-uploader": "0.5.0",
     "react-redux": "5.0.3",
     "react-router": "4.0.0",
@@ -40,7 +40,6 @@
     "react-select": "1.0.0-rc.3",
     "react-side-effect": "1.1.0",
     "react-svg-use": "2.1.0",
-    "react-webstorage": "0.0.4",
     "redux": "3.6.0"
   },
   "devDependencies": {
@@ -130,6 +129,7 @@
     "style-loader": "0.13.2",
     "svg-url-loader": "2.0.2",
     "uglify-js": "2.8.12",
+    "url-loader": "0.5.8",
     "webpack": "2.2.1",
     "webpack-chunk-hash": "0.4.0",
     "webpack-cleanup-plugin": "0.5.1",
