@@ -1,13 +1,14 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import css from './styles.scss'
 
 export default class Button extends Component {
 
     static propTypes = {
-        type: React.PropTypes.string,
-        label: React.PropTypes.string,
-        variant: React.PropTypes.string,
-        children: React.PropTypes.element.isRequired
+        type: PropTypes.string,
+        label: PropTypes.string,
+        variant: PropTypes.string,
+        children: PropTypes.element.isRequired
     }
 
     static defaultProps = {
