@@ -1,12 +1,10 @@
 import React from 'react'
-import { render } from 'react-dom'
 import App from 'router'
+import { render } from 'react-dom'
 
 // import { Provider } from 'react-redux'
 // import store from './store'
 // import router from './router'
-
-// require('es6-promise').polyfill()
 
 render(<App />, document.getElementById('root'))
 // render(<Provider store={store}>{router}</Provider>,, document.getElementById('root'))
